refactor: rename typings to @types, globals.d.ts to environment.d.ts
